+++ podcasts/Sharing/SharingView.swift
@@ -2,6 +2,12 @@ import SwiftUI
 import PocketCastsDataModel
 import PocketCastsUtils
 
+class ClipResult: ObservableObject {
+    @MainActor @Published var progress: Float = 0
+    @Published var exportURL: URL?
+    @Published var croppedURL: URL?
+}
+
 class ClipTime: ObservableObject {
     @Published var start: TimeInterval
     @Published var end: TimeInterval
@@ -22,8 +28,6 @@ struct SharingView: View {
         var shareType: UTType? = nil
     }
 
-    @EnvironmentObject var theme: Theme
-
     private enum Constants {
         static let descriptionMaxWidth: CGFloat = 200
         static let tabViewPadding: CGFloat = 80 // A value which represents extra padding for UIPageControl of the TabView
@@ -35,6 +39,7 @@ struct SharingView: View {
     @State private var isExporting: Bool = false
 
     @ObservedObject var clipTime: ClipTime
+    @StateObject var clipResult = ClipResult()
 
     init(destinations: [ShareDestination], selectedOption: SharingModal.Option, selectedStyle: ShareImageStyle = .large) {
         self.destinations = destinations
@@ -52,38 +57,7 @@ struct SharingView: View {
         VStack {
             title
             tabView
-            switch shareable.option {
-            case .episode, .podcast, .currentPosition:
-                buttons
-            case .clip(let episode, _):
-                VStack(spacing: 12) {
-                    MediaTrimBar(clipTime: clipTime, episode: episode)
-                        .frame(height: 72)
-                        .tint(color)
-                    HStack {
-                        Text(L10n.clipStartLabel(TimeFormatter.shared.playTimeFormat(time: clipTime.start)))
-                        Spacer()
-                        Text(L10n.clipDurationLabel(TimeFormatter.shared.playTimeFormat(time: clipTime.end - clipTime.start)))
-                    }
-                    .foregroundStyle(.white.opacity(0.5))
-                    .font(.caption.weight(.semibold))
-                    Button(L10n.clip, action: {
-                        withAnimation {
-                            shareable.option = .clipShare(episode, clipTime, shareable.style, Progress())
-                            isExporting = true
-                        }
-                    }).buttonStyle(RoundedButtonStyle(theme: theme, backgroundColor: color))
-                }
-                .padding(.horizontal, 16)
-            case .clipShare(_, _, _, let progress):
-                if progress.fileURL == nil && progress.fractionCompleted != 1 {
-                    ProgressView(progress)
-                        .tint(color)
-                        .padding()
-                } else {
-                    buttons
-                }
-            }
+            SharingFooterView(clipTime: clipTime, option: $shareable.option, isExporting: $isExporting, clipResult: clipResult, destinations: destinations, style: shareable.style)
         }
         .task(id: isExporting, {
             await exportIfNeeded()
@@ -92,22 +66,11 @@ struct SharingView: View {
         .foregroundStyle(Color.white)
     }
 
-    var color: Color {
-        switch shareable.option {
-        case .clip(let episode, _), .clipShare(let episode, _, _, _):
-            PlayerColorHelper.backgroundColor(for: episode)?.color ?? PlayerColorHelper.playerBackgroundColor01(for: theme.activeTheme).color
-        default:
-            PlayerColorHelper.playerBackgroundColor01(for: theme.activeTheme).color
-        }
-    }
-
     @ViewBuilder var title: some View {
         VStack {
-            Text(shareable.option.shareTitle)
+            Text(shareable.option.shareTitle(style: shareable.style))
                 .font(.headline)
             switch shareable.option {
-            case .clip:
-                EmptyView() // Don't show the description to give extra space for trim view
             case .clipShare(let episode, let clipTime, _, _):
                 Button(action: {
                     isExporting = false
@@ -125,12 +88,13 @@ struct SharingView: View {
                 }
                 .padding(.top, 14)
             default:
-                Text(L10n.shareDescription)
-                    .font(.subheadline)
-                    .foregroundStyle(.secondary)
-                    .multilineTextAlignment(.center)
-                    .frame(maxWidth: Constants.descriptionMaxWidth)
+                EmptyView()
             }
+            Text(shareable.style.shareDescription(option: shareable.option) ?? "‏")
+                .font(.subheadline)
+                .foregroundStyle(.secondary)
+                .multilineTextAlignment(.center)
+                .frame(maxWidth: Constants.descriptionMaxWidth)
         }
     }
 
@@ -140,8 +104,13 @@ struct SharingView: View {
                 switch shareable.option {
                 case .clipShare(_, _, let style, _):
                     image(style: style, containerHeight: proxy.size.height)
-                default:
+                case .clip:
                     ForEach(ShareImageStyle.allCases, id: \.self) { style in
+                        image(style: style, containerHeight: proxy.size.height)
+                    }
+                default:
+                    let styles = ShareImageStyle.allCases.filter { $0 != .audio }
+                    ForEach(styles, id: \.self) { style in
                         image(style: style, containerHeight: proxy.size.height)
                     }
                 }
@@ -152,41 +121,12 @@ struct SharingView: View {
 
     @ViewBuilder func image(style: ShareImageStyle, containerHeight: CGFloat) -> some View {
         ShareImageView(info: shareable.option.imageInfo, style: style, angle: .constant(0))
+            .frame(width: style.previewSize.width, height: style.previewSize.height)
+            .fixedSize()
             .clipShape(RoundedRectangle(cornerRadius: 12))
             .tabItem { Text(style.tabString) }
             .id(style)
-			.scaleEffect((containerHeight - Constants.tabViewPadding) / ShareImageStyle.large.videoSize.height)
-    }
-
-    @ViewBuilder var buttons: some View {
-        HStack(spacing: 24) {
-            ForEach(destinations, id: \.self) { destination in
-                if let action = destination.action {
-                    button(destination: destination, style: shareable.style, action: action)
-                } else {
-                    if #available(iOS 16, *) {
-                        shareLink(option: shareable.option, destination: destination, style: shareable.style)
-                    }
-                }
-            }
-        }
-    }
-
-    @ViewBuilder func button(destination: ShareDestination, style: ShareImageStyle, action: @escaping ((SharingModal.Option, ShareImageStyle) -> Void)) -> some View {
-        Button {
-            action(shareable.option, shareable.style)
-        } label: {
-            view(for: destination)
-        }
-    }
-
-    @available(iOS 16.0, *)
-    @ViewBuilder func shareLink(option: SharingModal.Option, destination: ShareDestination, style: ShareImageStyle) -> some View {
-        ShareLink(items: shareItems(style: style), preview: { _ in
-            SharePreview(option.imageInfo.title, image: Image(uiImage: ShareImageView(info: option.imageInfo, style: style, angle: .constant(0)).snapshot()))
-        }) {
-            view(for: destination)
-        }
+            .scaleEffect((containerHeight - Constants.tabViewPadding) / ShareImageStyle.large.previewSize.height)
     }
 
     private func shareItems(style: ShareImageStyle) -> [Shareable] {
@@ -198,30 +138,13 @@ struct SharingView: View {
         return [media, (style != .audio ? image : nil)].compactMap { $0 }
     }
 
-    @ViewBuilder func view(for destination: ShareDestination) -> some View {
-        VStack {
-            destination.icon
-                .renderingMode(.template)
-                .font(size: 20, style: .body, weight: .bold)
-                .frame(width: 24, height: 24)
-                .padding(15)
-                .background {
-                    Circle()
-                        .foregroundStyle(.white.opacity(0.1))
-                }
-            Text(destination.name)
-                .font(.caption)
-                .foregroundStyle(.secondary)
-        }
-    }
-
     enum VideoExportError: Error {
         case failedToDownload
     }
 
     private func exportIfNeeded() async {
         guard isExporting,
-              case let .clipShare(episode, clipTime, shareable.style, progress) = shareable.option
+              case let .clipShare(episode, clipTime, shareable.style, _) = shareable.option
         else {
             return
         }
@@ -229,21 +152,35 @@ struct SharingView: View {
             isExporting = false
         }
         do {
-            let url = try await export(info: shareable.option.imageInfo,
+            try await Task.sleep(nanoseconds: 500_000_000) // Delay by half a second to let the UI update
+            clipResult.progress = Float.leastNormalMagnitude
+            let progress = Progress()
+            let observation = progress.observe(\.completedUnitCount) { progress, change in
+                Task.detached { @MainActor in
+                    clipResult.progress = Float(progress.completedUnitCount) / Float(progress.totalUnitCount)
+                }
+            }
+
+            let (fullURL, croppedURL) = try await export(info: shareable.option.imageInfo,
                                             style: shareable.style,
                                             episode: episode,
                                             startTime: CMTime(seconds: clipTime.start, preferredTimescale: 600),
                                             duration: CMTime(seconds: clipTime.end - clipTime.start, preferredTimescale: 600),
                                             progress: progress
             )
-            progress.fileURL = url
-            shareable.option = .clipShare(episode, clipTime, shareable.style, progress)
+
+            clipResult.exportURL = fullURL
+            clipResult.croppedURL = croppedURL
+
+            shareable.option = .clipShare(episode, clipTime, shareable.style, clipResult)
+
+            progress.completedUnitCount = progress.totalUnitCount
         } catch let error {
             FileLog.shared.addMessage("Failed Clip Export: \(error)")
         }
     }
 
-    private func export(info: ShareImageInfo, style: ShareImageStyle, episode: some BaseEpisode, startTime: CMTime, duration: CMTime, progress: Progress) async throws -> URL {
+    private func export(info: ShareImageInfo, style: ShareImageStyle, episode: some BaseEpisode, startTime: CMTime, duration: CMTime, progress: Progress) async throws -> (URL, URL?) {
         guard let playerItem = DownloadManager.shared.downloadParallelToStream(of: episode) else {
             throw VideoExportError.failedToDownload
         }
@@ -252,7 +189,7 @@ struct SharingView: View {
         case .audio:
             let url = FileManager.default.temporaryDirectory.appendingPathComponent("audio_export-\(UUID().uuidString)", conformingTo: .m4a)
             try await AudioClipExporter.exportAudioClip(from: playerItem.asset, startTime: startTime, duration: duration, to: url, progress: progress)
-            return url
+            return (url, nil)
         default:
             let size = CGSize(width: style.videoSize.width * 2, height: style.videoSize.height * 2)
             guard #available(iOS 16, *) else { // iOS 15 support will be added in a separate PR just to keep the line count down
@@ -261,69 +198,11 @@ struct SharingView: View {
             let parameters = VideoExporter.Parameters(duration: CMTimeGetSeconds(duration), size: size, episodeAsset: playerItem.asset, audioStartTime: startTime, audioDuration: duration, fileType: .mp4)
             let url = FileManager.default.temporaryDirectory.appendingPathComponent("video_export-\(UUID().uuidString)", conformingTo: .mpeg4Movie)
             try await VideoExporter.export(view: AnimatedShareImageView(info: info, style: style), with: parameters, to: url, progress: progress)
-            return url
-        }
-    }
-}
 
-@available(iOS 16.0, *)
-extension SharingView.Shareable: Transferable {
-    static var transferRepresentation: some TransferRepresentation {
-        DataRepresentation<Self>(exportedContentType: .m4a) { shareable in
-            switch shareable.option {
-            case .clipShare(_, _, _, let progress):
-                let fileURL = try progress.fileURL.throwOnNil()
-                return try Data(contentsOf: fileURL)
-            default:
-                assertionFailure("This should never run due to exporting conditions below")
-                throw Optional<Void>.OptionalNil()
-            }
-        }.exportingCondition { shareable in
-            guard shareable.shareType == .audio,
-                  case .clipShare = shareable.option else { return false }
-            switch shareable.option {
-            case .clipShare:
-                return true
-            default:
-                return false
-            }
-        }
-        .suggestedFileName("clip")
-        FileRepresentation<Self>(exportedContentType: .mpeg4Movie) { shareable in
-            switch shareable.option {
-            case .clipShare(_, _, _, let progress):
-                let fileURL = try progress.fileURL.throwOnNil()
-                return SentTransferredFile(fileURL)
-            default:
-                assertionFailure("This should never run due to exporting conditions below")
-                throw Optional<Void>.OptionalNil()
-            }
-        }.exportingCondition { shareable in
-            guard shareable.shareType == .video,
-                  case .clipShare = shareable.option else { return false }
-            switch shareable.option {
-            case .clipShare:
-                return true
-            default:
-                return false
-            }
-        }
-        DataRepresentation<Self>(exportedContentType: .png) { shareable in
-            return try await ShareImageView(info: shareable.option.imageInfo, style: shareable.style, angle: .constant(0)).snapshot().pngData().throwOnNil()
-        }.exportingCondition { shareable in
-            guard shareable.shareType == .image else { return false }
-            switch shareable.option {
-            case .clipShare:
-                return false
-            default:
-                return true
-            }
-        }
-        ProxyRepresentation { shareable in
-            try URL(string: shareable.option.shareURL).throwOnNil()
-        }
-        .exportingCondition { shareable in
-            return shareable.shareType == .url
+            let fittedSize = CGSize(width: shareable.style.videoSize.width * 2, height: shareable.style.videoSize.height * 2).fitting(aspectRatio: shareable.style.previewSize)
+            let croppedURL = try await AVAsset(url: url).crop(to: fittedSize)
+
+            return (url, croppedURL)
         }
     }
 }